# Faster 'uniform' for Char