Benchmark:

```
How many random numbers can we generate in a second on one thread?
  First, timing System.Random.next:
  Second, timing System.Random.random at different types:
      1,540,246 randoms generated [System.Random Chars]       ~ 1,125 cycles/int     ~ 51.64 cycles/int
  Next timing range-restricted System.Random.randomR:
      1,621,726 randoms generated [System.Random Chars]       ~ 1,069 cycles/int     ~ 921 cycles/int
Now 8 threads, reporting mean randoms-per-second-per-thread:
  First, timing System.Random.next:
  Second, timing System.Random.random at different types:
        614,347 randoms generated [System.Random Chars]       ~ 2,821 cycles/int     ~ 113 cycles/int
  Next timing range-restricted System.Random.randomR:
        708,875 randoms generated [System.Random Chars]       ~ 2,445 cycles/int     ~ 2,173 cycles/int
```